Data de Funcionario funcionando

+++ src/java/dao/FuncionarioDAO.java
@@ -4,8 +4,12 @@ import entidades.Funcionario;
 import java.util.List;
 
 public interface FuncionarioDAO {
+
     public void save(Funcionario f);
+
     public void delete(Funcionario f);
-    public Funcionario find(Long id);
+
+    public Funcionario find(int id);
+
     public List<Funcionario> all();
 }
